update harbor conformance test workflow
Revise the flow to test the latest master code rather than the demo site, as the demo site is not
stable enough, the testing namespace can be removed by any of player.

Also, the demo site is a fixed build, the fix of conformance test failuse will only be in the master code.

BTW, the flow is built for http as the conformance-test docker image has no way to inject the self-signed crt,
which created in the phase of install Harbor.

Signed-off-by: wang yan <wangyan@vmware.com>